refactor(core): inline portable shell scanner
