--- packages/shell-scan/package.json
@@ -1,31 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/package.json",
-  "name": "@opencode-ai/shell-scan",
-  "version": "0.0.0",
-  "type": "module",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/anomalyco/opencode.git",
-    "directory": "packages/shell-scan"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "files": [
-    "dist"
-  ],
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "build": "bun run script/build.ts",
-    "test": "bun test --only-failures",
-    "typecheck": "tsgo --noEmit"
-  },
-  "devDependencies": {
-    "@tsconfig/bun": "catalog:",
-    "@types/bun": "catalog:",
-    "@typescript/native-preview": "catalog:"
-  }
-}
