--- packages/shell-scan/script/build.ts
@@ -1,10 +1,0 @@
-#!/usr/bin/env bun
-
-import { $ } from "bun"
-import { rm } from "node:fs/promises"
-import { fileURLToPath } from "node:url"
-
-process.chdir(fileURLToPath(new URL("..", import.meta.url)))
-
-await rm("dist", { recursive: true, force: true })
-await $`bun tsc -p tsconfig.build.json`
