--- packages/shell-scan/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "@tsconfig/bun/tsconfig.json",
-  "compilerOptions": {
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "allowImportingTsExtensions": false,
-    "allowJs": false,
-    "noUncheckedIndexedAccess": false
-  },
-  "include": ["src", "test", "bench", "research"]
-}
